Broke up reshape_data function into multiple utils files
